requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
Django==4.0.7
django-appconf==1.0.5
django-cors-headers==3.13.0
django-jazzmin==2.5.0
django-read-only-admin==0.17.5
django-readonly-model==1.2
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.1
mysql-connector-python==8.0.30
mysqlclient==2.1.1
numpy==1.23.4
oauthlib==3.2.2
pandas==1.5.1
Pillow==9.3.0
protobuf==3.20.1
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.2.1
reportlab==3.6.12
requests==2.28.1
requests-oauthlib==1.3.1
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
threadpoolctl==3.1.0
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
